Get a `start_section` command implemented for data generation
